updates
level two draw goal first so that the grapple cooldown will show on top on the goal. and change sequence of code so that the color is correct.
updates in some numbers in movement.c so that the player cannot drop through the base platform. and change cooldown bar color. Changed some number for speed of getting pulled and some logic change so that the gravity is not more than player y speed when grappling